Recompiled with namespace and js translations added to student view.
