+++ group_project_v2/project_navigator.py
@@ -1,9 +1,11 @@
 """
 This module contains Project Navigator XBlock and it's children view XBlocks
 """
+import pkg_resources
 import logging
 from lazy.lazy import lazy
 from opaque_keys import InvalidKeyError
+from django import utils
 from xblock.core import XBlock
 from xblock.exceptions import NoSuchUsage
 from web_fragments.fragment import Fragment
@@ -178,6 +180,7 @@ class GroupProjectNavigatorXBlock(
         return validation
 
 
+@XBlock.needs("i18n")
 class ProjectNavigatorViewXBlockBase(
     CompletionMixin,
     XBlockWithPreviewMixin,
@@ -242,12 +245,36 @@ class ProjectNavigatorViewXBlockBase(
     def is_view_available(self):  # pylint: disable=no-self-use
         return True
 
+    @staticmethod
+    def resource_string(path):
+        """Handy helper for getting resources."""
+        data = pkg_resources.resource_string(__name__, path)
+        return data.decode("utf8")
+
+    def get_translation_content(self):
+        """
+        Returns JS content containing translations for user's language.
+        """
+        try:
+            return self.resource_string('public/js/translations/{lang}/textjs.js'.format(
+                lang=utils.translation.to_locale(utils.translation.get_language()),
+            ))
+        except IOError:
+            return self.resource_string('public/js/translations/en/textjs.js')
+
+        @property
+        def i18n_service(self):
+            """ Obtains translation service """
+            return self.runtime.service(self, "i18n")
+
     def render_student_view(self, context, add_resources_from=None):
         """
         Common code to render student view
         """
         fragment = Fragment()
-        fragment.add_content(loader.render_template(self.TEMPLATE_BASE + self.template, context))
+        fragment.add_content(loader.render_django_template(self.TEMPLATE_BASE + self.template,
+                                                           context=context,
+                                                           i18n_service=self.i18n_service))
 
         if self.css_file:
             add_resource(self, 'css', self.CSS_BASE + self.css_file, fragment)
@@ -264,7 +291,7 @@ class ProjectNavigatorViewXBlockBase(
         if add_resources_from:
             for frag in add_resources_from:
                 fragment.add_fragment_resources(frag)
-
+        fragment.add_javascript(self.get_translation_content())
         return fragment
 
     def author_view(self, _context):
