+++ group_project_v2/public/js/translations/ar/textjs.js
@@ -1,4 +1,8 @@
 
+            (function(global){
+                var GroupProjectV2XBlockI18N = {
+                  init: function() {
+                    
 
 (function(globals) {
 
@@ -154,3 +158,10 @@
 
 }(this));
 
+
+                  }
+                };
+                GroupProjectV2XBlockI18N.init();
+                global.GroupProjectV2XBlockI18N = GroupProjectV2XBlockI18N;
+            }(this));
+
