--- group_project_v2/public/js/translations/de-de/textjs.js
@@ -1,144 +1,0 @@
-
-
-(function(globals) {
-
-  var django = globals.django || (globals.django = {});
-
-  
-  django.pluralidx = function(count) { return (count == 1) ? 0 : 1; };
-  
-
-  /* gettext library */
-
-  django.catalog = django.catalog || {};
-  
-  var newcatalog = {
-    " Technical details: 403 error.": "Technische Details: 403 Fehler.", 
-    " Technical details: CSRF verification failed.": "Technische Details: CSRF-Verifizierung fehlgeschlagen.", 
-    "An error occurred while uploading your file. Please refresh the page and try again. If it still does not upload, please contact your Course TA.": "W\u00e4hrend dem Upload Ihrer Datei ist ein Fehler aufgetreten. Bitte aktualisieren Sie die Seite und versuchen Sie es erneut. Funktioniert der Upload immer noch nicht, wenden Sie sich bitte an Ihren Kurs-TA.", 
-    "Error": "Fehler", 
-    "Error refreshing statuses": "Fehler bei der Aktualisierung des Status", 
-    "Please select Group to review": "Bitte w\u00e4hlen Sie eine Gruppe zur \u00dcberpr\u00fcfung aus", 
-    "Please select Teammate to review": "Bitte w\u00e4hlen Sie Teammate zur \u00dcberpr\u00fcfung aus", 
-    "Resubmit": "Erneut einreichen", 
-    "Submit": "Einreichen", 
-    "Thanks for your feedback!": "Danke f\u00fcr Ihr Feedback!", 
-    "This task has been marked as complete.": "Diese Aufgabe wurde als abgeschlossen markiert.", 
-    "Upload cancelled by user.": "Upload von Nutzer abgebrochen.", 
-    "Upload cancelled.": "Upload abgebrochen.", 
-    "We encountered an error loading your feedback.": "Beim Laden Ihres Feedbacks ist ein Fehler aufgetreten.", 
-    "We encountered an error saving your feedback.": "Beim Speichern Ihres Feedbacks ist ein Fehler aufgetreten.", 
-    "We encountered an error saving your progress.": "Beim Speichern Ihres Fortschritts ist ein Fehler aufgetreten.", 
-    "We encountered an error.": "Ein Fehler ist aufgetreten."
-  };
-  for (var key in newcatalog) {
-    django.catalog[key] = newcatalog[key];
-  }
-  
-
-  if (!django.jsi18n_initialized) {
-    django.gettext = function(msgid) {
-      var value = django.catalog[msgid];
-      if (typeof(value) == 'undefined') {
-        return msgid;
-      } else {
-        return (typeof(value) == 'string') ? value : value[0];
-      }
-    };
-
-    django.ngettext = function(singular, plural, count) {
-      var value = django.catalog[singular];
-      if (typeof(value) == 'undefined') {
-        return (count == 1) ? singular : plural;
-      } else {
-        return value[django.pluralidx(count)];
-      }
-    };
-
-    django.gettext_noop = function(msgid) { return msgid; };
-
-    django.pgettext = function(context, msgid) {
-      var value = django.gettext(context + '\x04' + msgid);
-      if (value.indexOf('\x04') != -1) {
-        value = msgid;
-      }
-      return value;
-    };
-
-    django.npgettext = function(context, singular, plural, count) {
-      var value = django.ngettext(context + '\x04' + singular, context + '\x04' + plural, count);
-      if (value.indexOf('\x04') != -1) {
-        value = django.ngettext(singular, plural, count);
-      }
-      return value;
-    };
-
-    django.interpolate = function(fmt, obj, named) {
-      if (named) {
-        return fmt.replace(/%\(\w+\)s/g, function(match){return String(obj[match.slice(2,-2)])});
-      } else {
-        return fmt.replace(/%s/g, function(match){return String(obj.shift())});
-      }
-    };
-
-
-    /* formatting library */
-
-    django.formats = {
-    "DATETIME_FORMAT": "j. F Y H:i", 
-    "DATETIME_INPUT_FORMATS": [
-      "%d.%m.%Y %H:%M:%S", 
-      "%d.%m.%Y %H:%M:%S.%f", 
-      "%d.%m.%Y %H:%M", 
-      "%d.%m.%Y", 
-      "%Y-%m-%d %H:%M:%S", 
-      "%Y-%m-%d %H:%M:%S.%f", 
-      "%Y-%m-%d %H:%M", 
-      "%Y-%m-%d"
-    ], 
-    "DATE_FORMAT": "j. F Y", 
-    "DATE_INPUT_FORMATS": [
-      "%d.%m.%Y", 
-      "%d.%m.%y", 
-      "%Y-%m-%d"
-    ], 
-    "DECIMAL_SEPARATOR": ",", 
-    "FIRST_DAY_OF_WEEK": "1", 
-    "MONTH_DAY_FORMAT": "j. F", 
-    "NUMBER_GROUPING": "3", 
-    "SHORT_DATETIME_FORMAT": "d.m.Y H:i", 
-    "SHORT_DATE_FORMAT": "d.m.Y", 
-    "THOUSAND_SEPARATOR": ".", 
-    "TIME_FORMAT": "H:i", 
-    "TIME_INPUT_FORMATS": [
-      "%H:%M:%S", 
-      "%H:%M:%S.%f", 
-      "%H:%M"
-    ], 
-    "YEAR_MONTH_FORMAT": "F Y"
-  };
-
-    django.get_format = function(format_type) {
-      var value = django.formats[format_type];
-      if (typeof(value) == 'undefined') {
-        return format_type;
-      } else {
-        return value;
-      }
-    };
-
-    /* add to global namespace */
-    globals.pluralidx = django.pluralidx;
-    globals.gettext = django.gettext;
-    globals.ngettext = django.ngettext;
-    globals.gettext_noop = django.gettext_noop;
-    globals.pgettext = django.pgettext;
-    globals.npgettext = django.npgettext;
-    globals.interpolate = django.interpolate;
-    globals.get_format = django.get_format;
-
-    django.jsi18n_initialized = true;
-  }
-
-}(this));
-
