+++ group_project_v2/public/js/translations/zh_CN/textjs.js
@@ -16,6 +16,29 @@
 
   django.catalog = django.catalog || {};
   
+  var newcatalog = {
+    " Technical details: 403 error.": " \u6280\u672f\u8be6\u7ec6\u4fe1\u606f\uff1a403 \u9519\u8bef\u3002", 
+    " Technical details: CSRF verification failed.": " \u6280\u672f\u8be6\u7ec6\u4fe1\u606f\uff1aCSRF \u9a8c\u8bc1\u5931\u8d25\u3002", 
+    "An error occurred while uploading your file. Please refresh the page and try again. If it still does not upload, please contact your Course TA.": "\u4e0a\u4f20\u60a8\u7684\u6587\u4ef6\u65f6\u51fa\u9519\u3002\u8bf7\u5237\u65b0\u9875\u9762\u5e76\u91cd\u8bd5\u3002\u5982\u679c\u4ecd\u7136\u65e0\u6cd5\u4e0a\u4f20\uff0c\u8bf7\u8054\u7cfb\u60a8\u7684\u8bfe\u7a0b\u52a9\u6559\u3002", 
+    "Error": "\u9519\u8bef", 
+    "Error refreshing statuses": "\u5237\u65b0\u72b6\u6001\u65f6\u51fa\u9519", 
+    "Please select Group to review": "\u8bf7\u9009\u62e9\u5c0f\u7ec4\u4ee5\u8fdb\u884c\u5ba1\u67e5", 
+    "Please select Teammate to review": "\u8bf7\u9009\u62e9\u7ec4\u5458\u4ee5\u8fdb\u884c\u5ba1\u67e5", 
+    "Resubmit": "\u91cd\u65b0\u63d0\u4ea4", 
+    "Submit": "\u63d0\u4ea4", 
+    "Thanks for your feedback!": "\u611f\u8c22\u60a8\u7684\u53cd\u9988\uff01", 
+    "This task has been marked as complete.": "\u6b64\u4efb\u52a1\u5df2\u6807\u8bb0\u4e3a\u5b8c\u6210\u3002", 
+    "Upload cancelled by user.": "\u7528\u6237\u5df2\u53d6\u6d88\u4e0a\u4f20\u3002", 
+    "Upload cancelled.": "\u4e0a\u4f20\u5df2\u53d6\u6d88\u3002", 
+    "We encountered an error loading your feedback.": "\u52a0\u8f7d\u60a8\u7684\u53cd\u9988\u65f6\u9047\u5230\u9519\u8bef\u3002", 
+    "We encountered an error saving your feedback.": "\u4fdd\u5b58\u60a8\u7684\u53cd\u9988\u65f6\u9047\u5230\u9519\u8bef\u3002", 
+    "We encountered an error saving your progress.": "\u4fdd\u5b58\u60a8\u7684\u8fdb\u5ea6\u65f6\u9047\u5230\u9519\u8bef\u3002", 
+    "We encountered an error.": "\u6211\u4eec\u9047\u5230\u4e86\u9519\u8bef\u3002"
+  };
+  for (var key in newcatalog) {
+    django.catalog[key] = newcatalog[key];
+  }
+  
 
   if (!django.jsi18n_initialized) {
     django.gettext = function(msgid) {
